:truck: Move `Favorites` to a separate file
